Refactor the 'add a placement' journey spec
Ahead of refactoring the implementation code for this journey, I've
decided to tweak the spec slightly so it's less dependent on the
existing implementation.

Changes I've made:

- Refactor the 'cancel' link scenario. Despite being a valuable test, it
  was slightly unrealistic because it depended on navigating directly to
each step and clicking 'Cancel'. It also depended on each step rendering
even if it wasn't supposed to be part of that journey.  The refactor
covers the same behaviour, but drives the form as a user would rather
than navigating directly using Rails URL helper methods.

- Remove the scenario "when I tamper with the form URL". This was
  another test which was slightly unrealistic. While I think it's still
valuable to test these sorts of edge cases, we had to do some contrived
stuff to achieve it in this spec. I think it would probably be better
covered in a request spec. It was also dependent on the specific
implementation, so it wouldn't work with the wizard refactor.

- Document a bug with a "TODO". Both the current implementation and the
  test contain a bug where they don't show the 'year group' step, when
they actually should. This will be fixed by the wizard refactor, at
which point we'll need to update the spec to test the correct behaviour.

- Remove an unused private method when_i_visit_the_check_your_answers_page

- Use the visually hidden text in the 'Change' links rather than
  depending on hardcoded page URLs which would break under a different
implementation.