Replace JS filters with a reusable implementation
This introduces a "filter-search" Stimulus JS controller which is more
reusable than the implementation it replaces.

The refactor has also put a focus on progressive enhancement by changing
the way the search input is created. It's now generated and injected
into the page by JavaScript, which means it won't appear for browsers
which don't have JavaScript enabled.

This behaviour is preferable because the filter relies on JavaScript to
work – therefore, it wouldn't make sense for a search input to exist on
the page without it working.

This simplifies the implementation slightly by reducing the need for
matching pairs of data attributes to be applied to legends and list
elements.

Instead, the markup required is simplified just to applying the
controller to the containing "app-filter__option" div of each filter
where a search is needed.