Update Claims filters to use the new JS controller
Update the Claims search filter markup it works with the updated
"filter-search" Stimulus JS controller.

We have deliberately removed one specific check from the system spec.
The previous behaviour was that the filter search input's value would
persist across page loads. For example, typing into the filter search,
choosing an option, and clicking "Apply filters" would cause the page to
reload and the search input to re-populate with the previous search
term. However I don't think this tends to be the expected or desired
behaviour with interactions like this. It's also not how the new
"filter-search" JS controller behaves, so we've removed this test.

Functionality-wise, everything else is unchanged.

I also noticed that the "View claims" system spec was running slowly. It
turns out the spec was asserting that some elements _don't_ exist on the
page. But since they don't exist on the page, Capybara was waiting a
couple of seconds in case they appear. In this case I've disabled the
wait functionality because we already know the page has finished
loading, since the previous assertion would have passed. This cuts about
2 seconds off from the runtime of this spec.