Fix readability of hidden text in 'change' links
'Change' links contain visually hidden text to help screenreader users
know what they're going to change.

Previously, the change links contained "Add" in the step names. For
example, the change link for mentors read "Change Add mentors". Which is
a bit confusing.

I've updated this so it now reads more naturally. For example, the
change mentors link will now read "Change Mentors".

Note: I'm not too worried about the capitalisation of the step name in
this case. I don't think it makes a difference to the way screenreaders
read it.