Create a new controller for 'Add placement' wizard
This controller implements the new 'Add placement' wizard
`Placements::AddPlacementWizard` and updates existing links to the
'build' controller to use this instead.

I decided to go ahead and create a new controller rather than try to
re-use the existing controller. It felt easier to start from a clean
slate and build it up rather than refactor what was already there.

I've also updated the 'Add placement' button on the placements index
page so it now links to this new controller rather than the (now unused)
'build' controller.

This should be an invisible change for end users, as demonstrated by the
system spec which still passes under the new controller and wizard
implementation. The only thing to have changes is the URLs of pages, but
this is unavoidable and doesn't form part of our system specs (nor do we
expect our users to be wedded to these existing URLs).