Introduce a wizard pattern for multi-step forms
This commit introduces a new 'wizard' class to help with buidling
multi-step forms. This will help us to ask [one thing per page][1], as
advocated for by the GOV.UK Service Manual and [described][2] in the
GOV.UK Design System documentation.

Out of the box, Ruby on Rails is tailored towards 'RESTful' user
interfaces which deal with creating objects in their entirety during one
single POST request. This makes for a slightly awkward fit with the
'one thing per page' pattern in government digital services.

`Placements::BaseWizard` and `Placements::BaseStep` intend to provide an
opinionated implementation of how multi-step forms should be implemented
within this application. It draws inspiration from the [dfe-wizard][3]
gem, although does not aim to recreate it. While dfe-wizard provides a
solid structure for building multi-step forms, it comes with its own set
of opinions which are incompatible with our app. For example, its
'store' concept seems to be designed for saving partial objects into the
database as ActiveRecord objects – whereas we've decided to store the
partially complete state of a wizard into the user's session. Our wizard
pattern actually avoids the responsibility of producing the resultant
object (usually an ActiveRecord object) altogether – making it possible
to implement wizards that don't create objects at all.

I should also give an honourable mention to the [govuk-wizardry][4] gem.
It's for similar reasons that we decided not to use this – it comes with
a set of opinions which don't easily fit with our existing application.

This change was very much borne out of a refactor of our existing
implementation rather than looking to create something new or reinvent
the wheel. It's a distillation of what was already happening within our
app.

I've included a mock wizard called `BurgerOrderWizard` so we can unit
test the `BaseWizard` and `BaseStep` classes without tying them to a
specific implementation of a wizard within our app. I've tried to keep
it as lightweight as possible while also demonstrating some of the
features available to wizard classes.

[1]: https://www.gov.uk/service-manual/design/form-structure#start-with-one-thing-per-page
[2]: https://design-system.service.gov.uk/patterns/question-pages/
[3]: https://github.com/DFE-Digital/dfe-wizard
[4]: https://github.com/DFE-Digital/govuk-wizardry