Build a wizard for the 'Add placement' journey
This is the first 'real' implementation of the wizard class.

The `Placements::AddPlacementWizard` and associated
`Placements::AddPlacementWizard::*Step` classes model the multi-page
flow for adding new placements in the application.

The 'Add placement' flow is currently implemented by the 'build'
controller `Placements::Schools::Placements::BuildController`. This
controller has already had a preparatory refactor to use 'step' classes,
but this commit introduces a wizard class for the journey.