Tighten up policy rules for placements
I've introduced a new permission called "add_placement_journey" to the
Placement policy.

This is the permission used by the 'Add placement' wizard controller,
reflecting the fact that the wizard itself should be seen as a
self-contained unit. Since it's not a traditional CRUD/resourceful
controller, we don't need to authenticate each method (new, edit,
update) of the wizard controller.

Hopefully this provides a clearer mapping between the policy and where
it's applied.

I've also DRYed up a duplicate rule in the policy, as well as the
associated spec file.