Remove seed step from the deployment workflow
Now that review apps are deployed using the command `rails db:prepare`,
we no longer need to explicitly seed the database when spinning up new
review apps.

This is because `rails db:prepare` will automatically seed the database
if it was empty to begin with.

This leads to more predictable behaviour in line with other
environments, where seeds only run once rather than once per deployment.