Use factories instead of Active Record fixtures
Use of Active Record fixtures is discouraged in our codebase. We prefer
to use FactoryBot factories to generate records in our tests.

Given this, the use of fixtures for 'claim window' and 'academic year'
records broke the principle of least astonishment.

I've removed these fixtures and have configured RSpec to create the
required records specifically for those tests that need them.