Add a check for pending GoodJob migrations
The GoodJob gem uses our app's existing Postgres database, and sometimes
it needs to run database migrations to update its tables.

This commit adds a check to the RSpec test suite so we know when GoodJob
has migrations that need to be run.

Without this, there's the chance our database silently falls out of sync
with GoodJob's required schema, leading to the potential for errors at
runtime. This is what happened when we tried updating from GoodJob v3 to
v4 without realising we needed to run migrations. This test should catch
future instances of this problem before we can merge and deploy the
change.