Implement multi-tab journeys for edit school contact wizard
https://trello.com/c/05kqUr3N/831-spike-out-how-we-could-support-multi-tab-wizard-journeys

Adds the required changes to support unique session states between edit school contact journeys and creates a new multi-tab system spec to cover the new functionality. Unique session states are achieved by passing a state key via the route, which allows us to find the correct state from the users session. Not that this does not allow a user to share their state with other users as this data is still stored in their local cache.