Move logic closer to the code that needs it
I've relocated the logic that was originally defined as
`Claims::Claim::MentorTrainingForm#max_hours_equals_maximum_claimable_hours?`
because it was only needed and used in the component
`Claims::Claim::MentorTrainingForm::DisclaimerComponent`.

The component spec already covered the required behaviour, so I was able
to simply move the code and cut out an unnecessary layer of abstraction.