Make `custom_hours_selected?` method private
This method doesn't need to be part of the form object's public
interface, so I've made it private. It's only used by the validation
rules inside this class.

I've also dropped the memoization of this method. It's a simple and
cheap enough method to run that I don't think it's harmful to run it
twice. It simplifies the code and doesn't add any significant overhead.