Stop assuming we're in the current academic year
As alluded to [in this comment][1], an assumption was being made that
the training allowance should always be calculated against the _current_
academic year.

While this is correct at the moment, because claims can only be created
and edited in the current academic year, it's still an uncomfortable
coupling.

A more sensible coupling would be to use the academic year associated
with the claim in scope of the class. That's a more sensible source of
truth for the academic year of the claim.

[1]: https://github.com/DFE-Digital/itt-mentor-services/pull/1062#discussion_r1790093806