Only run seeds in development environments
This drastically speeds up the provisioning of databases in test
environments (e.g. CI).

Prior to this, the test environment would download a full copy of GIAS
(among other things!) and load it into the test database, only for it to
be discarded since our test suite runs with truncation enabled.

This should be a safe change because these database seeds are only
intended for use in development environments anyway. They're of no use
to us in test or production.