Fix a flaky system spec
This system spec was unpredictably passing and failing.

It looks like this was caused by pulling the 'last' record from the
database, but not providing a sort order. You can't rely on the order of
records in the database unless you specify one – especially since we use
UUID strings as our primary keys.