Add refresher training to Mentor Training model
This change fully integrates refresher training into the Mentor Training
model.

1. Make the 'hours_completed' validation dynamic

Use the new 'training allowance' calculation to specify a dynamic
maximum for the field, rather than it being hardcoded to an upper bound
of 20 hours.

2. Set the 'training_type' field

So the calculated training type gets persisted to the database. Before
this, it would always be saved as 'initial' training. I've applied this
as a before_validation hook so the training type can still be validated
against the enum defined in this model.