Fix a bug which broke no-JS organisation search
There are three forms in the app which feature an autocomplete component
to help users search for and select an organisation (i.e. a school or
provider). These forms have fallback implementations for when JavaScript
is not working on the user's device.

We identified a bug whereby users on the no-JS journey could not change
their search term after having selected an organisation.

This was happening due to there being two competing 'sources of truth'
for what the user's search term was:

1. The original search term is stored in the "id" attribute on the
initial "OrganisationStep" (or "SchoolStep") step. Note that the naming
of this attribute is slightly misleading – but in the case of JS working
correctly, this attribute truly will hold the ID of the selected
organsiation. Without JS, it holds the search term as typed into the
text input field.

2. The search results step also holds a copy of the search term. This
was populated by first pulling in the search term from the previous
state, but then persisting its value within its own state.

The existence of the second 'source of truth' meant the original step
could never effectively update the search term, because the search
results step had already 'cached' its own copy of the attribute itself.

To resolve this bug, I've removed the internal 'cache' of that value
from the search results step. Instead it'll always refer to the initial
'organisation' step as the source of truth for the user's search term.